Refactor code from redundant crud place to router.
