+++ server/backend/api/app/crud/crud_forest_client.py
@@ -1,8 +1,5 @@
-import json
 import logging
-from typing import List
 
-from api.app.integration.forest_client.forest_client import ForestClient
 from api.app.models import model as models
 from sqlalchemy.orm import Session
 
@@ -60,23 +57,3 @@ def find_or_create(db: Session, forest_client_number: str, requester: str):
 
     LOGGER.debug(f"Forest_Client {fam_forest_client.client_number_id} found.")
     return fam_forest_client
-
-
-def search(p_client_number: str) -> List[schemas.FamForestClient]:
-    LOGGER.debug(f"Forest Client - 'search' with parameter: {p_client_number}.")
-    fc_api = ForestClient()
-    fc_json_list = fc_api.find_by_client_number(p_client_number)  # json object List
-    results = list(map(__map_api_results, fc_json_list))
-    LOGGER.debug(f"Result: {results}")
-    return results
-
-
-def __map_api_results(item) -> schemas.FamForestClient:
-    """
-    Private method to map api result to schemas.FamForestClient
-    """
-    parsed = json.loads(
-        json.dumps(item),  # need json string format, so dumps from 'dic' type 'item'.
-        object_hook=schemas.FamForestClient.from_api_json
-    )
-    return parsed
